+++ lib/mocktail/handles_dry_call/fulfills_stubbing.rb
@@ -1,15 +1,19 @@
 require_relative "fulfills_stubbing/finds_satisfaction"
+require_relative "fulfills_stubbing/describes_unsatisfied_stubbing"
 
 module Mocktail
   class FulfillsStubbing
     def initialize
       @finds_satisfaction = FindsSatisfaction.new
+      @describes_unsatisfied_stubbing = DescribesUnsatisfiedStubbing.new
     end
 
     def fulfill(dry_call)
       if (stubbing = satisfaction(dry_call))
         stubbing.satisfied!
         stubbing.effect&.call(dry_call)
+      else
+        StubReturnedNil.new(@describes_unsatisfied_stubbing.describe(dry_call))
       end
     end