+++ lib/mocktail/imitates_type/makes_double/gathers_fakeable_instance_methods.rb
@@ -1,0 +1,21 @@
+module Mocktail
+  class GathersFakeableInstanceMethods
+    def gather(type)
+      methods = type.instance_methods + [
+        (:respond_to_missing? if type.private_method_defined?(:respond_to_missing?))
+      ].compact
+
+      methods.reject { |m|
+        ignore?(type, m)
+      }
+    end
+
+    def ignore?(type, method_name)
+      ignored_ancestors.include?(type.instance_method(method_name).owner)
+    end
+
+    def ignored_ancestors
+      Object.ancestors
+    end
+  end
+end