+++ lib/mocktail/share/stringifies_method_name.rb
@@ -1,0 +1,11 @@
+module Mocktail
+  class StringifiesMethodName
+    def stringify(call)
+      [
+        call.original_type.name,
+        call.singleton ? "." : "#",
+        call.method
+      ].join
+    end
+  end
+end