+++ lib/mocktail/value/double.rb
@@ -1,11 +1,10 @@
 module Mocktail
-  class Double
-    attr_reader :original_type, :dry_type, :dry_instance
-
-    def initialize(original_type:, dry_type:, dry_instance:)
-      @original_type = original_type
-      @dry_type = dry_type
-      @dry_instance = dry_instance
-    end
+  class Double < Struct.new(
+    :original_type,
+    :dry_type,
+    :dry_instance,
+    :dry_methods,
+    keyword_init: true
+  )
   end
 end