+++ lib/mocktail/value/double_data.rb
@@ -1,0 +1,10 @@
+module Mocktail
+  class DoubleData < Struct.new(
+    :type,
+    :double,
+    :calls,
+    :stubbings,
+    keyword_init: true
+  )
+  end
+end