+++ lib/mocktail/value/type_replacement_data.rb
@@ -1,0 +1,13 @@
+module Mocktail
+  class TypeReplacementData < Struct.new(
+    :type,
+    :replaced_method_names,
+    :calls,
+    :stubbings,
+    keyword_init: true
+  )
+    def double
+      type
+    end
+  end
+end