+++ lib/mocktail/value/unsatisfied_stubbing.rb
@@ -1,0 +1,8 @@
+module Mocktail
+  class UnsatisfiedStubbing < Struct.new(
+    :call,
+    :other_stubbings,
+    keyword_init: true
+  )
+  end
+end